requirements.txt
flask
flask_restful
flask_sqlalchemy
PyJWT
pytest
